refactor: update import in tests after merge
